test: move sea script to test/fixtures
Refs: https://github.com/nodejs/node/pull/45038#discussion_r1101361319
